requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
asyncer==0.0.15
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chevron==0.14.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cuid==0.4
dataclasses-json==0.6.7
Deprecated==1.3.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.129.0
filelock==3.24.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
instructor==1.14.5
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.14
langchain-experimental==0.4.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langsmith==0.7.6
Lazify==0.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.16.0
nest-asyncio==1.6.0
numpy==2.4.2
openai==2.21.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b1
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.5.2
propcache==0.4.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.22
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
sse-starlette==3.2.0
starlette==0.52.1
streamlit==1.54.0
syncer==2.0.3
tabulate==0.9.0
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tomli==2.4.0
tornado==6.5.4
tqdm==4.67.3
typer==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
watchdog==6.0.0
watchfiles==1.1.1
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
